Add cartographer, which can calculate number of agents in each catchment area

+++ src/main/scala/org/ifc/figssamel/drc/location/GPSCoordinates.scala
@@ -19,4 +19,4 @@ case class GPSCoordinates(lat: Lat, lng: Lng) extends Ordered[GPSCoordinates] {
   }
 }
 
-case class CatchmentArea(squareEdgeLengthMeters: Double, numberWithinArea: Double)
+case class CatchmentArea(squareEdgeLengthMeters: Double, numberWithinArea: Long)
